// src/lib/types/openstreetmap.ts
export type NominatimData = {
  addresstype: string;
  boundingbox: [string, string, string, string];
  class: string;
  display_name: string;
  importance: number;
  lat: string;
  lon: string;
  name: string;
  osm_id: number;
  osm_type: string;
  place_id: string;
  place_rank: number;
  type: string;
};